+++ apps/cart/src/middleware.ts
@@ -87,7 +87,8 @@ export const config = {
 		 * - sitemap (sitemap file)
 		 * - site.webmanifest (site.webmanifest file)
 		 * - robots (robots file)
+		 * - png files
 		 */
-		'/((?!api|_next|_static|.well-known|favicon|logos|sitemap|site.webmanifest|robots).*)',
+		'/((?!api|_next|_static|.well-known|favicon|logos|sitemap|site.webmanifest|robots|.*\\.png).*)',
 	],
 };
