+++ packages/ui/elements/mdx-editor/index.tsx
@@ -28,7 +28,6 @@ const Editor = dynamic(
 				<Skeleton className='h-8 w-full' />
 				<Skeleton className='h-8 w-full' />
 				<Skeleton className='h-8 w-full' />
-				{/* <Skeleton className='h-8 w-full' /> */}
 			</div>
 		),
 	},
